Refactor Periscope modules into smaller files
